Revise the process of rotating before shooting in attacking logic.
The following applies to both rotating to ball and rotating to goal:
If the robot has not started rotating {
	if the ball/goal is not seen, search for it;
	if the ball/goal is seen but the bearing is small enough, skip rotating;
	if the ball/goal is seen and there is significant bearing, start rotating
}
if the robot has started rotating, wait until finished;
if the robot has finished rotating, go to next state.

Head mode for rotating to goal is also revised (first use the head mode when the goal is last found; switch to FarMid after rotating body).